+++ doc/changelog.md
@@ -64,6 +64,28 @@
 * Add support for ``var`` of general observables
   [#124](https://github.com/PennyLaneAI/catalyst/pull/124)
 
+* Add support for Jacobian product operations. The function ``jvp(f, params, tangents)`` returns
+  the Jacobian-vector product of a function `f`, and the ``vjp(f, params, cotangents)`` returns
+  its vector-Jacobian product.
+
+  [#98](https://github.com/PennyLaneAI/catalyst/pull/98)
+
+  An example of a newly supported workflow:
+
+
+  ``` python
+  from catalyst import jvp
+
+  def f(p):
+    return jnp.stack([1*p, 2*p, 3*p])
+
+  @qjit
+  def workflow(params, tangent):
+      return jvp(f, [params], [tangent])
+
+  workflow(jnp.zeros([4], dtype=float), jnp.ones([4], dtype=float))
+  ```
+
 <h3>Improvements</h3>
 
 * Improving error handling by throwing descriptive and unified expressions for runtime
@@ -113,9 +135,10 @@ This release contains contributions from (in alphabetical order):
 
 Ali Asadi,
 David Ittah,
+Erick Ochoa Lopez,
 Jacob Mai Peng,
 Romain Moyard,
-Erick Ochoa Lopez.
+Sergei Mironov.
 
 # Release 0.1.2